- Changed userid to uuid - Added store for uuid:username - Changed remove user method to uuid - Added ROADMAP.md

+++ src/modules/users/repositories/users.repository.ts
@@ -2,6 +2,12 @@ import { Injectable } from '@nestjs/common';
 import { RedisService } from '@/common/libs/redis/redis.service';
 import { UserEntity } from '@/modules/users/entities/user.entity';
 
+
+enum REDIS_KEYS {
+    USERNAME = 'username',
+    UUID = 'uuid',
+}
+
 @Injectable()
 export class UsersRepository {
     constructor(
@@ -10,19 +16,32 @@ export class UsersRepository {
     
     async create(user: UserEntity): Promise<UserEntity> {
         await this.redisService.client.hSet(
-            `user:${user.username}`,
+            `${REDIS_KEYS.USERNAME}:${user.username}`,
             'password',
             user.password,
         );
+        
         await this.redisService.client.hSet(
-            `user:${user.username}`,
+            `${REDIS_KEYS.USERNAME}:${user.username}`,
             'id',
-            user.id,
+            user.uuid,
         );
+        
+        await this.redisService.client.hSet(
+            `${REDIS_KEYS.UUID}:${user.uuid}`,
+            'username',
+            user.username,
+        );
+        
         return user;
     }
     
     async clear(): Promise<void> {
+        await this.removeKeys(REDIS_KEYS.USERNAME);
+        await this.removeKeys(REDIS_KEYS.UUID);
+    }
+    
+    private async removeKeys(key: REDIS_KEYS): Promise<void> {
         let cursor = '0';
         let keysToDelete: string[] = [];
         
@@ -33,7 +52,7 @@ export class UsersRepository {
             } = await this.redisService.client.scan(
                 cursor,
                 {
-                    MATCH: 'user:*',
+                    MATCH: `${key}:*`,
                     COUNT: 100,
                 },
             );
@@ -45,16 +64,23 @@ export class UsersRepository {
             await this.redisService.client.del(keysToDelete);
         }
     }
-    
-    async remove(username: string): Promise<void> {
+   
+    async remove(userUuid: string): Promise<void> {
+        const user = await this.getByUuid(userUuid);
+        if (!user) {
+            return;
+        }
         await this.redisService.client.del(
-            `user:${username}`,
+            `${REDIS_KEYS.UUID}:${user.uuid}`,
+        );
+        await this.redisService.client.del(
+            `${REDIS_KEYS.USERNAME}:${user.username}`,
         );
     }
     
     async getByUsername(username: string): Promise<UserEntity | null> {
         const exist = await this.redisService.client.exists(
-            `user:${ username }`,
+            `${REDIS_KEYS.USERNAME}:${ username }`,
         );
         
         if (!exist) {
@@ -62,12 +88,40 @@ export class UsersRepository {
         }
         
         const user = await this.redisService.client.hGetAll(
-            `user:${ username }`,
+            `${REDIS_KEYS.USERNAME}:${ username }`,
         );
         
         return new UserEntity(
-            user.id,
+            user.uuid,
             username,
+            user.password,
+        );
+    }
+    
+    async getByUuid(userUuid: string): Promise<UserEntity | null> {
+        const userByUuid = await this.redisService.client.hGetAll(
+            `${REDIS_KEYS.UUID}:${userUuid}`,
+        )
+        
+        if (!userByUuid) {
+            return null;
+        }
+        
+        const exist = await this.redisService.client.exists(
+            `${REDIS_KEYS.USERNAME}:${ userByUuid.username }`,
+        );
+        
+        if (!exist) {
+            return null;
+        }
+        
+        const user = await this.redisService.client.hGetAll(
+            `${REDIS_KEYS.USERNAME}:${ userByUuid.username }`,
+        );
+        
+        return new UserEntity(
+            user.uuid,
+            userByUuid.username,
             user.password,
         );
     }
